create separate post process for resolving
